Atualizacao automática

+++ src/main/java/br/org/coletivoJava/integracoes/restMavmail/api/pojo/ClienteMav.java
@@ -165,7 +165,7 @@ public class ClienteMav implements ItfPessoa {
     }
 
     @Override
-    public int configIDPeloNome() {
+    public Long configIDPeloNome() {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
